setup popToRoot in WPF, pull out CustomViewPresenter and MvxPanelPopToRootPresentationHint in UWP
