+++ XPlatformMenus/XPlatformMenus.UWP/Views/MvxPanelPopToRootPresentationHint.cs
@@ -1,0 +1,9 @@
+﻿using MvvmCross.Core.ViewModels;
+
+namespace XPlatformMenus.UWP.Views
+{
+    public class MvxPanelPopToRootPresentationHint : MvxPresentationHint
+    {
+
+    }
+}
