+++ XPlatformMenus/XPlatformMenus.WPF/Views/MvxPanelPopToRootPresentationHint.cs
@@ -1,0 +1,9 @@
+﻿using MvvmCross.Core.ViewModels;
+
+namespace XPlatformMenus.WPF.Views
+{
+    public class MvxPanelPopToRootPresentationHint : MvxPresentationHint
+    {
+
+    }
+}
