+++ XPlatformMenus/XPlatformMenus.WPF/XPlatformMenus.WPF.csproj
@@ -76,6 +76,7 @@
     <Compile Include="Setup.cs" />
     <Compile Include="Views\Base\BaseView.cs" />
     <Compile Include="Views\Base\BasePage.cs" />
+    <Compile Include="Views\CustomViewPresenter.cs" />
     <Compile Include="Views\Help\HelpView.xaml.cs">
       <DependentUpon>HelpView.xaml</DependentUpon>
     </Compile>
@@ -94,6 +95,7 @@
     <Compile Include="Views\Menu\MenuView.xaml.cs">
       <DependentUpon>MenuView.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Views\MvxPanelPopToRootPresentationHint.cs" />
     <Compile Include="Views\MvxRegionAttribute.cs" />
     <Compile Include="Views\MvxMultiRegionWpfViewPresenter.cs" />
     <Compile Include="Views\MvxWpfExtensionMethods.cs" />
